+++ Sources/macimed/main.swift
@@ -8,10 +8,6 @@ do {
    try Sock.startDaemon()
 } catch let e as AppError {
    Sock.log(e.message)
-} catch let e as IMEError {
-   Sock.log(e.message)
-} catch let e as CmdError {
-   Sock.log(e.message)
 } catch {
    Sock.log("Unexpected error\n")
 }